refactor: decompose MongoDB backend
